requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
python-dotenv>=1.0.0

aiofiles>=23.0.0

# Data sources
pytrends>=4.9.2
praw>=7.7.1
feedparser>=6.0.11
requests>=2.31.0

# LLM (use whichever you have keys for)
openai>=1.30.0
anthropic>=0.26.0

# Testing
pytest>=8.0.0
